backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Pillow==10.0.0
python-dotenv==1.0.0
email-validator==2.0.0
phonenumbers==8.13.17